@@ -30,13 +30,14 @@
     "@ngx-translate/core": "^13.0.0",
     "@ngx-translate/http-loader": "^6.0.0",
     "ajv": "^6.12.6",
-    "autoprefixer": "^10.4.12",
+    "autoprefixer": "^10.4.13",
     "firebase": "^9.8.2",
     "moment": "^2.29.1",
     "ngx-cookie-service": "^14.0.0",
+    "post": "^0.0.1",
     "postcss": "^8.4.16",
     "rxjs": "^7.5.5",
-    "tailwindcss": "^3.1.8",
+    "tailwindcss": "3.1.8",
     "tslib": "^2.0.0",
     "uuid": "^8.3.2",
     "zone.js": "~0.11.4"
@@ -69,6 +70,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "^4.6.4"
+    "typescript": "4.7.4"
   }
 }
