@@ -2,7 +2,7 @@ import { Congregation } from 'src/app/_interfaces/congregation.interface';
 import { Site } from 'src/app/_interfaces/site.interface';
 import { User, UserKey } from 'src/app/_interfaces/user.interface';
 import { PersonalShifts } from '../_interfaces/personal-shifts.interface';
-import { ShiftHours } from '../_interfaces/shift-hours.interface';
+import { ShiftHour } from '../_interfaces/shift-hours.interface';
 import { Shift } from '../_interfaces/shift.interface';
 import { SiteShifts } from '../_interfaces/site-shifts.interface';
 import { Statistic } from '../_interfaces/statistic.interface';
@@ -24,13 +24,23 @@ export interface ApiInterface {
   updateUserActivation: (
     uuid: string,
     activate: boolean
-  ) => Promise<{ date: string; hour: ShiftHours; site: Site }[]>;
+  ) => Promise<{ date: string; hour: ShiftHour; site: Site }[]>;
 
   readCongregations: () => Promise<Congregation[]>;
+  createCongregation: (cong: Omit<Congregation, 'uuid' | 'activate' | 'order'>) => Promise<void>
+  updateCongregation: (cong: Omit<Congregation, 'activate' | 'order'>) => Promise<void>
+  changeCongregationActivation: (cong: Congregation) => Promise<void>
 
   readSites: () => Promise<Site[]>;
+  createSite: (site: Omit<Site, 'uuid'>) => Promise<void>;
+  updateSite: (site: Site) => Promise<void>
+  changeSiteActivation: (site: Site) => Promise<void>
 
-  readShiftHoursList: () => Promise<ShiftHours[]>;
+  readShiftHours: () => Promise<ShiftHour[]>;
+  createShiftHour: (shifthour: Omit<ShiftHour, 'uuid' | 'activate' | 'deliver'>) => Promise<void>
+  updateShiftHour: (shiftHour: ShiftHour) => Promise<void>
+  changeShiftHourActivation: (shifthour: ShiftHour) => Promise<void>
+  changeShiftHourDelivery: (shifthour: ShiftHour) => Promise<void>
 
   readShiftsByMonth: (yearMonth: string) => Promise<Shift[]>; // yyyy-MM
   readShiftsByDate: (date: string) => Promise<Shift[]>; // yyyy-MM-dd
