@@ -1,7 +1,10 @@
-import { AfterViewInit, Component, Inject, OnInit } from '@angular/core';
+import { AfterViewInit, Component, Inject, OnInit, OnDestroy } from '@angular/core';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
 import { MatDialogRef, MAT_DIALOG_DATA } from '@angular/material/dialog';
+import { Observable, Subject } from 'rxjs';
 import { filter, first } from 'rxjs/operators';
+import { Permission } from 'src/app/_enums/permission.enum';
+import { Statistic } from 'src/app/_interfaces/statistic.interface';
 import { AuthorityService } from 'src/app/_services/authority.service';
 import { StatisticsService } from 'src/app/_services/statistics.service';
 
@@ -10,9 +13,11 @@ import { StatisticsService } from 'src/app/_services/statistics.service';
   templateUrl: './statistic-editor.component.html',
   styleUrls: ['./statistic-editor.component.scss']
 })
-export class StatisticEditorComponent implements OnInit, AfterViewInit {
+export class StatisticEditorComponent implements OnInit, OnDestroy {
 
   statisticForm!: FormGroup;
+  managerAccess$!: Observable<boolean>;
+  destroy$ = new Subject<void>();
 
   constructor(
     private dialogRef: MatDialogRef<StatisticEditorComponent>,
@@ -25,17 +30,15 @@ export class StatisticEditorComponent implements OnInit, AfterViewInit {
   ngOnInit(): void {
     const disabled = this.data.mode === 'view';
     this.statisticForm = this.formBuilder.group({
-      attendance: [{ value: '', disabled}, [Validators.required, Validators.pattern(/[0-9]+/)]],
-      tracts: [{ value: '', disabled}, [Validators.required, Validators.pattern(/[0-9]+/)]],
-      scriptures: [{ value: '', disabled}, [Validators.required, Validators.pattern(/[0-9]+/)]],
-      videos: [{ value: '', disabled}, [Validators.required, Validators.pattern(/[0-9]+/)]],
-      acceptReturnVisit: [{ value: '', disabled}, [Validators.required, Validators.pattern(/[0-9]+/)]],
-      returnVisits: [{ value: '', disabled}, [Validators.required, Validators.pattern(/[0-9]+/)]],
-      experience: [{ value: '', disabled}],
+      attendance: [{ value: '', disabled}, [Validators.required, Validators.pattern(/^[0-9]+$/)]],
+      tracts: [{ value: '', disabled}, [Validators.required, Validators.pattern(/^[0-9]+$/)]],
+      videos: [{ value: '', disabled}, [Validators.required, Validators.pattern(/^[0-9]+$/)]],
+      returnVisits: [{ value: '', disabled}, [Validators.required, Validators.pattern(/^[0-9]+$/)]],
+      startingBibleStudies: [{ value: '', disabled}, [Validators.required, Validators.pattern(/^[0-9]+$/)]],
     });
-  }
 
-  ngAfterViewInit(): void {
+    this.managerAccess$ = this.authorityService.canAccess(Permission.MANAGER);
+
     if (this.data.mode !== 'create') {
       const yearMonth = this.data.date.slice(0, 7);
       this.statisticsService.getStatistic(yearMonth, this.data.uuid).pipe(
@@ -46,15 +49,23 @@ export class StatisticEditorComponent implements OnInit, AfterViewInit {
           this.statisticForm.setValue({
             attendance: _statistic.attendance,
             tracts: _statistic.tracts,
-            scriptures: _statistic.scriptures,
             videos: _statistic.videos,
-            acceptReturnVisit: _statistic.acceptReturnVisit,
             returnVisits: _statistic.returnVisits,
-            experience: _statistic.experience,
+            startingBibleStudies: _statistic.startingBibleStudies | 0,
           }, {emitEvent: false});
         }
       });
     }
+  }
+
+  ngOnDestroy(): void {
+    this.destroy$.next();
+    this.destroy$.complete();
+  }
+
+  onEditClick = () => {
+    this.data.mode = 'edit';
+    Object.values(this.statisticForm.controls).forEach(_control => _control.enable());
   }
 
   onCancelClick = () => {
@@ -63,7 +74,7 @@ export class StatisticEditorComponent implements OnInit, AfterViewInit {
 
   onSubmitClick = () => {
     if (this.statisticForm.valid) {
-      this.statisticsService.createStatistic({
+      const statistic: Statistic = {
         uuid: this.data.uuid,
         date: this.data.date,
         createdByUuid: this.authorityService.currentUserUuid$.value as string,
@@ -71,12 +82,17 @@ export class StatisticEditorComponent implements OnInit, AfterViewInit {
         activate: true,
         attendance: this.statisticForm.get('attendance')?.value,
         tracts: this.statisticForm.get('tracts')?.value,
-        scriptures: this.statisticForm.get('scriptures')?.value,
         videos: this.statisticForm.get('videos')?.value,
-        acceptReturnVisit: this.statisticForm.get('acceptReturnVisit')?.value,
         returnVisits: this.statisticForm.get('returnVisits')?.value,
-        experience: this.statisticForm.get('experience')?.value,
-      });
+        startingBibleStudies: this.statisticForm.get('startingBibleStudies')?.value,
+      }
+
+      if (this.data.mode === 'create') {
+        this.statisticsService.createStatistic(statistic);
+      } else { // this.data.mode === 'edit'
+        this.statisticsService.updateStatistic(statistic);
+      }
+
       this.dialogRef.close(true);
     } else {
       this.statisticForm.markAllAsTouched();
