@@ -1,8 +1,8 @@
-export interface ShiftHours {
+export interface ShiftHour {
   uuid: string;
   name: string;
   startTime: string; // HH:mm
   endTime: string; // HH:mm
   activate: boolean;
-  deliver:boolean;
+  deliver: boolean;
 }
