@@ -36,9 +36,19 @@ import { LoginDialogComponent } from './_elements/dialogs/login-dialog/login-dia
 import { StatisticEditorComponent } from './_elements/dialogs/statistic-editor/statistic-editor.component';
 import { ShiftCardComponent } from './_elements/shift-table/shift-card/shift-card.component';
 import { ShiftTableComponent } from './_elements/shift-table/shift-table.component';
+import { ArrayFilterPipe } from './_pipes/array-filter.pipe';
+import { SettingsComponent } from './settings/settings.component';
+import { SiteComponent } from './settings/sites/sites.component';
+import { ShfitHoursComponent } from './settings/shfitHours/shfitHours.component';
+import { CongregationsComponent } from './settings/congregations/congregations.component';
+import { SiteEditorComponent } from './_elements/dialogs/site-editor/site-editor.component';
+import { SiteCreatorComponent } from './_elements/dialogs/site-creator/site-creator.component';
+import { ShiftHoursCreatorComponent } from './_elements/dialogs/shiftHour-creator/shiftHour-creator.component';
+import { ShiftHoursEditorComponent } from './_elements/dialogs/shiftHour-editor/shiftHour-editor.component';
 import { YearMonthSelectComponent } from './_elements/year-month-select/year-month-select.component';
 import { LineBindingDialogComponent } from './line-binding/line-binding-dialog/line-binding-dialog.component';
-
+import { CongregationCreatorComponent } from './_elements/dialogs/congregation-creator/congregation-creator.component';
+import { CongregationEditorComponent } from './_elements/dialogs/congregation-editor/congregation-editor.component';
 // AoT requires an exported function for factories of translate module
 export function HttpLoaderFactory(http: HttpClient) {
   return new TranslateHttpLoader(http);
@@ -66,11 +76,23 @@ export function HttpLoaderFactory(http: HttpClient) {
     CalendarHeaderComponent,
     ForceRefreshDirective,
     HoursListComponent,
+    SettingsComponent,
+    SiteComponent,
+    ShfitHoursComponent,
+    CongregationsComponent,
+    SiteEditorComponent,
+    SiteCreatorComponent,
+    OpeningShiftsComponent,
+    YearMonthSelectComponent,
     OpeningShiftsComponent,
     YearMonthSelectComponent,
     LoginComponent,
     LineBindingComponent,
     LineBindingDialogComponent,
+    ShiftHoursCreatorComponent,
+    ShiftHoursEditorComponent,
+    CongregationCreatorComponent,
+    CongregationEditorComponent,
   ],
   imports: [
     BrowserModule,
@@ -106,4 +128,4 @@ export function HttpLoaderFactory(http: HttpClient) {
   ],
   bootstrap: [AppComponent],
 })
-export class AppModule {}
+export class AppModule { }
